finance_calculator.py

+++ python_introduction/finance_calculator.py
@@ -10,7 +10,7 @@ total_monthly_expenses = int(input("Enter your total monthly expenses:"))
 monthly_savings = monthly_income - total_monthly_expenses
 
 # Project Annual Savings:
-projected_savings = monthly_savings * 12 + monthly_savings * 12 *0.05
+projected_savings = int(monthly_savings * 12 + monthly_savings * 12 * 0.05)
 
 # Output Results:
 print("Your monthly savings are $",monthly_savings)
